Move function specific stuff into own file
